(BF4) Showing decimal K/D ratio in PlayerListWidget.

+++ src/client/game/bf4/PlayerListWidget.cpp
@@ -214,7 +214,7 @@ void PlayerListWidget::listPlayers(const QList<PlayerInfo> &playerList)
         playerItem->setText(4, QString::number(player.getScore()));
         playerItem->setText(5, QString::number(player.getPing()));
         playerItem->setText(6, player.getGuid());
-        playerItem->setText(7, QString::number(player.getDeaths() <= 0 ? player.getKills() : player.getKills() / player.getDeaths()));
+        playerItem->setText(7, QString::number(player.getDeaths() <= 0 ? (double) player.getKills() : (double) player.getKills() / (double) player.getDeaths()));
 
         // Add player item and team id to lists.
         playerItemList.insert(playerItem);
